+++ db/Economy.java
@@ -2,7 +2,7 @@ package org.example.listeners.db;
 
 public class Economy {
     private String user_id;
-    private double user_bal;
+    private long user_bal;
 
     public Economy() {
         // Default constructor
@@ -16,11 +16,11 @@ public class Economy {
         this.user_id = user_id;
     }
 
-    public double getUser_bal() {
+    public long getUser_bal() {
         return user_bal;
     }
 
-    public void setUser_bal(double user_bal) {
+    public void setUser_bal(long user_bal) {
         this.user_bal = user_bal;
     }
 }
